chore: move more files to utility
