+++ tests/dragonfly/management_test.py
@@ -1,5 +1,4 @@
 import pytest
-import asyncio
 from redis import asyncio as aioredis
 from redis.exceptions import ResponseError
