+++ tests/dragonfly/utility/seeder.py
@@ -1,6 +1,5 @@
 import itertools
 import logging
-import sys
 import asyncio
 from redis import asyncio as aioredis
 import redis
@@ -10,55 +9,8 @@ import itertools
 import time
 import difflib
 import json
-import subprocess
-import os
 from enum import Enum
-
-
-def chunked(n, iterable):
-    """Transform iterable into iterator of chunks of size n"""
-    it = iter(iterable)
-    while True:
-        chunk = tuple(itertools.islice(it, n))
-        if not chunk:
-            return
-        yield chunk
-
-
-def eprint(*args, **kwargs):
-    """Print to stderr"""
-    print(*args, file=sys.stderr, **kwargs)
-
-
-def gen_test_data(n, start=0, seed=None):
-    for i in range(start, n):
-        yield "k-" + str(i), "v-" + str(i) + ("-" + str(seed) if seed else "")
-
-
-def batch_fill_data(client, gen, batch_size=100):
-    for group in chunked(batch_size, gen):
-        client.mset({k: v for k, v, in group})
-
-
-async def wait_available_async(client: aioredis.Redis, timeout=10):
-    """Block until instance exits loading phase"""
-    its = 0
-    start = time.time()
-    while (time.time() - start) < timeout:
-        try:
-            await client.get("key")
-            return
-        except aioredis.ResponseError as e:
-            if "MOVED" in str(e):
-                # MOVED means we *can* serve traffic, but 'key' does not belong to an owned slot
-                return
-            assert "Can not execute during LOADING" in str(e)
-
-        # Print W to indicate test is waiting for replica
-        print("W", end="", flush=True)
-        await asyncio.sleep(0.01)
-        its += 1
-    raise RuntimeError("Client did not become available in time!")
+from .utility import eprint
 
 
 class SizeChange(Enum):
@@ -584,49 +536,3 @@ class DflySeederFactory:
 
     def create(self, **kwargs):
         return DflySeeder(log_file=self.log_file, **kwargs)
-
-
-async def disconnect_clients(*clients):
-    await asyncio.gather(*(c.connection_pool.disconnect() for c in clients))
-
-
-def gen_ca_cert(ca_key_path, ca_cert_path):
-    # We first need to generate the tls certificates to be used by the server
-
-    # Generate CA (certificate authority) key and self-signed certificate
-    # In production, CA should be generated by a third party authority
-    # Expires in one day and is not encrtypted (-nodes)
-    # X.509 format for the key
-    step = rf'openssl req -x509 -newkey rsa:4096 -days 1 -nodes -keyout {ca_key_path} -out {ca_cert_path} -subj "/C=GR/ST=SKG/L=Thessaloniki/O=KK/OU=AcmeStudios/CN=Gr/emailAddress=acme@gmail.com"'
-    subprocess.run(step, shell=True)
-
-
-def gen_certificate(
-    ca_key_path, ca_certificate_path, certificate_request_path, private_key_path, certificate_path
-):
-    # Generate Dragonfly's private key and certificate signing request (CSR)
-    step1 = rf'openssl req -newkey rsa:4096 -nodes -keyout {private_key_path} -out {certificate_request_path} -subj "/C=GR/ST=SKG/L=Thessaloniki/O=KK/OU=Comp/CN=Gr/emailAddress=does_not_exist@gmail.com"'
-    subprocess.run(step1, shell=True)
-
-    # Use CA's private key to sign dragonfly's CSR and get back the signed certificate
-    step2 = rf"openssl x509 -req -in {certificate_request_path} -days 1 -CA {ca_certificate_path} -CAkey {ca_key_path} -CAcreateserial -out {certificate_path}"
-    subprocess.run(step2, shell=True)
-
-
-class EnvironCntx:
-    def __init__(self, **kwargs):
-        self.updates = kwargs
-        self.undo = {}
-
-    def __enter__(self):
-        for k, v in self.updates.items():
-            if k in os.environ:
-                self.undo[k] = os.environ[k]
-            os.environ[k] = v
-
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        for k, v in self.updates.items():
-            if k in self.undo:
-                os.environ[k] = self.undo[k]
-            else:
-                del os.environ[k]
